+++ backend/app/middleware/__init__.py
@@ -1,0 +1,2 @@
+# SPDX-FileCopyrightText: 2026 Ivan Dodik
+# SPDX-License-Identifier: AGPL-3.0-only